Refactor source localization runner
